Feat: Added the Service Package @50%

+++ src/main/java/com/agroconnect/frogger/factory/DefaultOrderFactory.java
@@ -4,11 +4,13 @@ import com.agroconnect.frogger.entity.Order;
 import com.agroconnect.frogger.entity.OrderStatus;
 import org.springframework.stereotype.Component;
 
+import java.math.BigInteger;
+
 @Component
 public class DefaultOrderFactory implements OrderFactory {
 
     @Override
-    public Order createOrder(int customerId, int productId, int quantity, double totalPrice, String status) {
+    public Order createOrder(BigInteger customerId, BigInteger productId, int quantity, double totalPrice, String status) {
         Order order = new Order();
         order.setCustomerId((long) customerId);
         order.setProductId((long) productId);
